PDFTemplateResponse.get_override_settings is easier to override.
Now, PDFTemplateResponse.default_override_settings exists to
dynamically generate the settings to override. This will be documented
in usage.rst.

To support this, pathname2fileurl() is smart enough to figure out
whether a pathname ought to be converted into a file:// URL. Unit
tests for this new implementation now exist.