utils.py: wkhtmltopdf now handles WKHTMLTOPDF_CMD_ARGS defaults.
Always prepend the default arguments to the command line for now.
Dealing with conflicts is an interesting problem, because now we need
to know which things should be clobbered, and which shouldn't.

I suspect that the proper answer is to leave out things from
WKHTMLTOPDF_CMD_ARGS that you would want to override.