Patch python3 / mod_wsgi incompatibility on sys.stderrr.fileno() call
Root cause of this issue is in python3 / mod_wsgi.  See:  https://github.com/GrahamDumpleton/mod_wsgi/issues/85

This patch detects if there is something screwy with fileno() and simply skips setting stderr if there is to avoid the AttributeError that occurs otherwise.   Pull https://github.com/incuna/django-wkhtmltopdf/pull/40 fixed this same issue when it arose for python 2.7, but the nature of the issue seems to have shifted, making it trickier to detect that the fileno() method cannot be called.