Extracted logic from views - fixed tests.
The logic was coupled with the views which made things difficult
to reuse if you wanted to use the pdf generation somehwere else
than in a view.

With this patch, the logic has been moved to `utils.py` and should
be more easy to reuse.

Tests have been adapted and made compatible with django > 1.7