Fix unidecode bytes error on python2
`bytes` takes no additional args in python2 unlike it's taking
an encoding in python3.

Unidecode returns a unicode string in python3 and a bytestring
in python2 which, I believe, was the main cause of error #71.

Now, regardless of whether unidecode is included, all strings
passing through http_quote will be encoded to ascii which should
fix both issues.

Also included is a fix for a failing test when unidecode is used.
Unidecode's `_unidecode` function ignores characters greater
than 0xefff, which '\xe2\x99\xa5' (the heart symbol) is. This
caused users with unidecode to fail '.pdf' was produced rather
than the expected '?.pdf'.